Refs #84170 CCA: Moved the collective.cover modifier to the extras.py file and added zcml condition for the modifier.
